+++ urc_bt_nodes/CMakeLists.txt
@@ -28,37 +28,22 @@ include_directories(
   include
 )
 
-add_library(bt_log_info SHARED src/base/log_info.cpp)
-ament_target_dependencies(bt_log_info ${dependencies})
-target_compile_definitions(bt_log_info PRIVATE BT_PLUGIN_EXPORT)
-
-add_library(bt_call_trigger SHARED src/actions/call_trigger.cpp)
-ament_target_dependencies(bt_call_trigger ${dependencies})
-target_compile_definitions(bt_call_trigger PRIVATE BT_PLUGIN_EXPORT)
-
-add_library(bt_call_generate_plan SHARED src/actions/call_generate_plan.cpp)
+add_library(bt_call_generate_plan SHARED src/call_generate_plan.cpp)
 target_compile_definitions(bt_call_generate_plan PRIVATE BT_PLUGIN_EXPORT)
 ament_target_dependencies(bt_call_generate_plan ${dependencies})
 
-add_library(bt_follow_path SHARED src/actions/follow_path.cpp)
-target_compile_definitions(bt_follow_path PRIVATE BT_PLUGIN_EXPORT)
-ament_target_dependencies(bt_follow_path ${dependencies})
+add_library(bt_call_follow_path SHARED src/call_follow_path.cpp)
+target_compile_definitions(bt_call_follow_path PRIVATE BT_PLUGIN_EXPORT)
+ament_target_dependencies(bt_call_follow_path ${dependencies})
 
-add_library(bt_status_light_publisher SHARED src/actions/status_light_publisher.cpp)
-target_compile_definitions(bt_status_light_publisher PRIVATE BT_PLUGIN_EXPORT)
-ament_target_dependencies(bt_status_light_publisher ${dependencies})
-
-add_library(bt_update_current_pose SHARED src/subscribers/update_current_pose.cpp)
-target_compile_definitions(bt_update_current_pose PRIVATE BT_PLUGIN_EXPORT)
-ament_target_dependencies(bt_update_current_pose ${dependencies})
+add_library(bt_call_search SHARED src/call_search.cpp)
+target_compile_definitions(bt_call_search PRIVATE BT_PLUGIN_EXPORT)
+ament_target_dependencies(bt_call_search ${dependencies})
 
 set(plugins
-  bt_log_info
-  bt_call_trigger
   bt_call_generate_plan
-  bt_follow_path
-  bt_status_light_publisher
-  bt_update_current_pose
+  bt_call_follow_path
+  bt_call_search
 )
 
 install(TARGETS
