+++ urc_bt_nodes/src/call_search.cpp
@@ -1,4 +1,4 @@
-#include "urc_bt_nodes/actions/search.hpp"
+#include "urc_bt_nodes/call_search.hpp"
 #include "behaviortree_cpp/basic_types.h"
 #include <nav_msgs/msg/detail/path__struct.hpp>
 #include <rclcpp/logging.hpp>
@@ -10,14 +10,14 @@ namespace behavior::actions
 bool Search::setGoal(Goal & goal)
 {
   uint16_t get_goal = std::stoi(getInput<std::string>("goal_type").value());
-  RCLCPP_INFO(node_->get_logger(), "Setting Goal...");
+  RCLCPP_INFO(logger(), "Setting Goal...");
   goal.goal_type = get_goal;
   return true;
 }
 
 void Search::onHalt()
 {
-  RCLCPP_INFO(node_->get_logger(), "Halting...");
+  RCLCPP_INFO(logger(), "Halting...");
 }     // namespace behavior::actions
 
 BT::NodeStatus Search::onFeedback(const std::shared_ptr<const Feedback> feedback)
@@ -28,7 +28,7 @@ BT::NodeStatus Search::onFeedback(const std::shared_ptr<const Feedback> feedback
 BT::NodeStatus Search::onResultReceived(const WrappedResult & wr)
 {
   RCLCPP_INFO(
-    node_->get_logger(), "Finished following path with error code: %hu.", wr.result->error_code);
+    logger(), "Finished following path with error code: %hu.", wr.result->error_code);
   if (wr.result->error_code == 0) {
     return BT::NodeStatus::SUCCESS;
   }
@@ -38,10 +38,10 @@ BT::NodeStatus Search::onResultReceived(const WrappedResult & wr)
 
 BT::NodeStatus Search::onFailure(BT::ActionNodeErrorCode error)
 {
-  RCLCPP_ERROR(node_->get_logger(), "%s: Failed with error %s", name().c_str(), toStr(error));
+  RCLCPP_ERROR(logger(), "%s: Failed with error %s", name().c_str(), toStr(error));
   return BT::NodeStatus::FAILURE;
 }
 }
 
 #include "behaviortree_ros2/plugins.hpp"
-CreateRosNodePlugin(urc_msgs::action::SearchAruco, "Search");
+CreateRosNodePlugin(behavior::actions::Search, "Search");
